requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
cffi==1.16.0
click==8.1.7
frozenlist==1.4.1
idna==3.7
mccabe==0.7.0
multidict==6.0.5
peewee==3.17.6
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
python-dotenv==1.0.1
yarl==1.9.4
